requirements (1).txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
albumentations>=1.4.6
opencv-python>=4.8.1
numpy>=1.24.0
pandas>=1.5.3
scikit-learn>=1.3.0
pyyaml>=6.0
tqdm>=4.66.1
gradio>=4.31.5
pillow>=10.0.0
roboflow>=1.1.29
pytorch-grad-cam>=1.5.1
